refactor(tests): extract shared test harness to tests/common
Deduplicate TestServer, start_server, and make_client that were
copy-pasted across head_bucket, copy_object, and presign test files.
